requirements.txt
Mastodon.py==1.8.1
confluent-kafka==2.5.3
python-dotenv==1.0.1
